# apps/api/openapi/v1.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Shelf API",
    "version": "1.0.0"
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer"
      },
      "cookieAuth": {
        "type": "apiKey",
        "in": "cookie",
        "name": "better-auth.session_token"
      }
    },
    "schemas": {
...
      "def-3": {
        "type": "object",
        "required": ["error"],
        "properties": {
          "error": {
            "type": "object",
            "required": ["code", "message", "retryable", "requestId"],
            "properties": {
              "code": {
                "type": "string",
                "enum": [
                  "AUTHENTICATION_REQUIRED",
                  "AUTHORIZATION_DENIED",
                  "INVALID_REQUEST",
                  "IDEMPOTENCY_CONFLICT",
                  "ARTIFACT_NOT_FOUND",
                  "REVISION_NOT_FOUND",
                  "SHARE_NOT_FOUND",
                  "ACCESS_CREDENTIAL_NOT_FOUND",
                  "RANGE_NOT_SATISFIABLE",
                  "MULTI_RANGE_UNSUPPORTED",
                  "REQUEST_CANCELLED",
                  "CONTENT_UNAVAILABLE",
                  "SERVICE_UNAVAILABLE",
                  "INTERNAL_ERROR"
                ],
                "title": "ErrorCode"
              },
              "message": {
                "type": "string",
                "minLength": 1,
                "maxLength": 512
              },
              "retryable": {
                "type": "boolean"
              },
              "requestId": {
                "type": "string",
                "minLength": 1,
                "maxLength": 128
              },
              "details": {
                "type": "array",
                "items": {
                  "type": "object",
                  "required": ["reason"],
                  "properties": {
                    "field": {
                      "type": "string",
                      "minLength": 1,
                      "maxLength": 128
                    },
                    "reason": {
                      "type": "string",
                      "minLength": 1,
                      "maxLength": 256
                    }
                  },
                  "additionalProperties": false,
                  "title": "ErrorDetail"
                },
                "maxItems": 32
              }
            },
            "additionalProperties": false
          }
        },
        "additionalProperties": false,
        "title": "ErrorEnvelope"
      },
...
    }
  },
  "paths": {
...
    "/api/v1/artifacts/{artifactId}/revisions": {
      "get": {
        "operationId": "listArtifactRevisionsV1",
        "summary": "List one artifact revision history newest first",
        "tags": ["artifacts"],
        "parameters": [
          {
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 100
            },
            "in": "query",
            "name": "limit",
            "required": false
          },
          {
            "schema": {
              "type": "string",
              "minLength": 1,
              "maxLength": 2048
            },
            "in": "query",
            "name": "cursor",
            "required": false
          },
          {
            "schema": {
              "type": "string",
              "pattern": "^art_[A-Za-z0-9_-]{22}$"
            },
            "in": "path",
            "name": "artifactId",
            "required": true,
            "description": "Opaque artifact identifier containing 128 bits of entropy."
          }
        ],
        "security": [
          {
            "bearerAuth": []
          },
          {
            "cookieAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": ["apiVersion", "artifactId", "workspaceId", "items", "nextCursor"],
                  "properties": {
                    "apiVersion": {
                      "type": "string",
                      "enum": ["v1"]
                    },
                    "artifactId": {
                      "type": "string",
                      "pattern": "^art_[A-Za-z0-9_-]{22}$",
                      "description": "Opaque artifact identifier containing 128 bits of entropy."
                    },
                    "workspaceId": {
                      "type": "string",
                      "minLength": 1,
                      "maxLength": 128
                    },
                    "items": {
                      "type": "array",
                      "items": {
                        "anyOf": [
                          {
                            "type": "object",
                            "required": [
                              "revisionId",
                              "revisionNumber",
                              "contentHash",
                              "createdAt",
                              "provenance",
                              "publisherMetadata",
                              "kind",
                              "originalFileName",
                              "mediaType",
                              "byteCount",
                              "fileCount",
                              "paths"
                            ],
                            "properties": {
                              "revisionId": {
                                "type": "string",
                                "pattern": "^rev_[A-Za-z0-9_-]{22}$",
                                "description": "Opaque revision identifier containing 128 bits of entropy."
                              },
                              "revisionNumber": {
                                "type": "integer",
                                "minimum": 1,
                                "maximum": 9007199254740991
                              },
                              "contentHash": {
                                "type": "string",
                                "pattern": "^sha256:[a-f0-9]{64}$"
                              },
                              "createdAt": {
                                "type": "string",
                                "pattern": "^\\d{4}-\\d{2}-\\d{2}T\\d{2}:\\d{2}:\\d{2}\\.\\d{3}Z$"
                              },
                              "provenance": {
                                "anyOf": [
                                  {
                                    "type": "object",
                                    "required": ["classification", "observed"],
                                    "properties": {
                                      "classification": {
                                        "type": "string",
                                        "enum": ["direct-publish"]
                                      },
                                      "observed": {
                                        "type": "object",
                                        "required": ["actorId", "operation"],
                                        "properties": {
                                          "actorId": {
                                            "type": "string",
                                            "minLength": 1,
                                            "maxLength": 128
                                          },
                                          "operation": {
                                            "type": "string",
                                            "enum": ["file.publish"]
                                          }
                                        },
                                        "additionalProperties": false
                                      }
                                    },
                                    "additionalProperties": false
                                  },
                                  {
                                    "type": "object",
                                    "required": ["classification", "observed", "source"],
                                    "properties": {
                                      "classification": {
                                        "type": "string",
                                        "enum": ["restore"]
                                      },
                                      "observed": {
                                        "type": "object",
                                        "required": ["actorId", "operation"],
                                        "properties": {
                                          "actorId": {
                                            "type": "string",
                                            "minLength": 1,
                                            "maxLength": 128
                                          },
                                          "operation": {
                                            "type": "string",
                                            "enum": ["revision.restore"]
                                          }
                                        },
                                        "additionalProperties": false
                                      },
                                      "source": {
                                        "type": "object",
                                        "required": ["revisionId"],
                                        "properties": {
                                          "revisionId": {
                                            "type": "string",
                                            "pattern": "^rev_[A-Za-z0-9_-]{22}$",
                                            "description": "Opaque revision identifier containing 128 bits of entropy."
                                          }
                                        },
                                        "additionalProperties": false
                                      }
                                    },
                                    "additionalProperties": false
                                  }
                                ]
                              },
                              "publisherMetadata": {
                                "type": "object",
                                "maxProperties": 32,
                                "additionalProperties": {
                                  "type": "string",
                                  "maxLength": 2048
                                }
                              },
                              "kind": {
                                "type": "string",
                                "enum": ["file"]
                              },
                              "originalFileName": {
                                "type": "string",
                                "minLength": 1,
                                "maxLength": 255
                              },
                              "mediaType": {
                                "type": "string",
                                "minLength": 1,
                                "maxLength": 255
                              },
                              "byteCount": {
                                "type": "integer",
                                "minimum": 1,
                                "maximum": 9007199254740991
                              },
                              "fileCount": {
                                "type": "number",
                                "enum": [1]
                              },
                              "paths": {
                                "type": "object",
                                "required": ["revision", "content"],
                                "properties": {
                                  "revision": {
                                    "type": "string",
                                    "pattern": "^/api/v1/revisions/[^/]+$"
                                  },
                                  "content": {
                                    "type": "string",
                                    "pattern": "^/api/v1/revisions/[^/]+/content$"
                                  }
                                },
                                "additionalProperties": false
                              }
                            },
                            "additionalProperties": false
                          },
                          {
                            "type": "object",
                            "required": [
                              "revisionId",
                              "revisionNumber",
                              "contentHash",
                              "createdAt",
                              "provenance",
                              "publisherMetadata",
                              "kind",
                              "rootName",
                              "byteCount",
                              "fileCount",
                              "paths"
                            ],
                            "properties": {
                              "revisionId": {
                                "type": "string",
                                "pattern": "^rev_[A-Za-z0-9_-]{22}$",
                                "description": "Opaque revision identifier containing 128 bits of entropy."
                              },
                              "revisionNumber": {
                                "type": "integer",
                                "minimum": 1,
                                "maximum": 9007199254740991
                              },
                              "contentHash": {
                                "type": "string",
                                "pattern": "^sha256:[a-f0-9]{64}$"
                              },
                              "createdAt": {
                                "type": "string",
                                "pattern": "^\\d{4}-\\d{2}-\\d{2}T\\d{2}:\\d{2}:\\d{2}\\.\\d{3}Z$"
                              },
                              "provenance": {
                                "anyOf": [
                                  {
                                    "type": "object",
                                    "required": ["classification", "observed"],
                                    "properties": {
                                      "classification": {
                                        "type": "string",
                                        "enum": ["direct-publish"]
                                      },
                                      "observed": {
                                        "type": "object",
                                        "required": ["actorId", "operation"],
                                        "properties": {
                                          "actorId": {
                                            "type": "string",
                                            "minLength": 1,
                                            "maxLength": 128
                                          },
                                          "operation": {
                                            "type": "string",
                                            "enum": ["file.publish"]
                                          }
                                        },
                                        "additionalProperties": false
                                      }
                                    },
                                    "additionalProperties": false
                                  },
                                  {
                                    "type": "object",
                                    "required": ["classification", "observed", "source"],
                                    "properties": {
                                      "classification": {
                                        "type": "string",
                                        "enum": ["restore"]
                                      },
                                      "observed": {
                                        "type": "object",
                                        "required": ["actorId", "operation"],
                                        "properties": {
                                          "actorId": {
                                            "type": "string",
                                            "minLength": 1,
                                            "maxLength": 128
                                          },
                                          "operation": {
                                            "type": "string",
                                            "enum": ["revision.restore"]
                                          }
                                        },
                                        "additionalProperties": false
                                      },
                                      "source": {
                                        "type": "object",
                                        "required": ["revisionId"],
                                        "properties": {
                                          "revisionId": {
                                            "type": "string",
                                            "pattern": "^rev_[A-Za-z0-9_-]{22}$",
                                            "description": "Opaque revision identifier containing 128 bits of entropy."
                                          }
                                        },
                                        "additionalProperties": false
                                      }
                                    },
                                    "additionalProperties": false
                                  }
                                ]
                              },
                              "publisherMetadata": {
                                "type": "object",
                                "maxProperties": 32,
                                "additionalProperties": {
                                  "type": "string",
                                  "maxLength": 2048
                                }
                              },
                              "kind": {
                                "type": "string",
                                "enum": ["folder"]
                              },
                              "rootName": {
                                "type": "string",
                                "minLength": 1,
                                "maxLength": 255
                              },
                              "byteCount": {
                                "type": "integer",
                                "minimum": 0,
                                "maximum": 9007199254740991
                              },
                              "fileCount": {
                                "type": "integer",
                                "minimum": 0,
                                "maximum": 1000
                              },
                              "paths": {
                                "type": "object",
                                "required": ["revision", "tree"],
                                "properties": {
                                  "revision": {
                                    "type": "string",
                                    "pattern": "^/api/v1/revisions/[^/]+$"
                                  },
                                  "tree": {
                                    "type": "string",
                                    "pattern": "^/api/v1/revisions/[^/]+/tree$"
                                  }
                                },
                                "additionalProperties": false
                              }
                            },
                            "additionalProperties": false
                          }
                        ]
                      },
                      "maxItems": 100
                    },
                    "nextCursor": {
                      "anyOf": [
                        {
                          "type": "string",
                          "minLength": 1,
                          "maxLength": 2048
                        },
                        {
                          "type": "null"
                        }
                      ]
                    }
                  },
                  "additionalProperties": false
                }
              }
            }
          },
          "400": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/def-3"
                }
              }
            }
          },
          "401": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/def-3"
                }
              }
            }
          },
          "403": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/def-3"
                }
              }
            }
          },
          "404": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/def-3"
                }
              }
            }
          },
          "500": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/def-3"
                }
              }
            }
          },
          "503": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/def-3"
                }
              }
            }
          }
        }
      }
    },
...
  }
}